Use GitHub test environments in e2e tests
